misc: add some meaningless tests
