+++ sennaar-rs/src/rossetta/mod.rs
@@ -3,3 +3,4 @@ pub mod clang_expr;
 pub mod clang_ty;
 pub mod clang_utils;
 pub mod to_registry;
+pub mod name_unnamed;
